Refactoring and Cleaning
