+++ ai/README.md
@@ -1,0 +1,35 @@
+# 🧠 AI Module — AI-Swings-Stocks
+
+This folder contains the **core AI system** powering the stock analysis and decision-making logic of [AI-Swings-Stocks](https://github.com/lakshaychhabra/AI-Swings-Stocks).  
+It includes autonomous agents, core trading logic, technical analysis tools, and LLM-based reasoning to analyze stocks for swing trading.
+
+---
+
+## 📌 Overview
+
+The AI system is built to analyze daily stock opportunities using multi-agent orchestration. It integrates:
+
+- **Technical analysis**
+- **Sentiment analysis** 
+- **Quantitative reasoning**
+- **Risk Assesment**
+- **Numbers of Stocks To trade** 
+- **Strategy Analysis**
+- **Ensemble engine**
+
+The AI logic is exposed via Python modules and optionally served via an API in `app.py`.
+
+---
+
+## ⚙️ Setup & Installation
+
+> Requires **Python 3.11+**
+
+```bash
+# Clone the repo
+git clone https://github.com/lakshaychhabra/AI-Swings-Stocks.git
+cd AI-Swings-Stocks/ai
+```
+
+# Install dependencies
+pip install -r requirements.txt
