+++ core/technical_analysis/README.md
