+++ core/technical_analysis/guide.md
