+++ core/technical_analysis/momentum.py
