+++ core/technical_analysis/others.py
