+++ core/technical_analysis/trend.py
